requirements_thu_am.txt
pyserial>=3.5
pyaudio>=0.2.11
pydub>=0.25.1
ffmpeg-python>=0.2.0